+++ gdm/cli.py
@@ -27,37 +27,38 @@ def main(args=None, function=None):
     project = argparse.ArgumentParser(add_help=False)
     project.add_argument('-r', '--root', metavar='PATH',
                          help="root directory of the project")
-    shared = {'formatter_class': common.WideHelpFormatter,
-              'parents': [project, debug]}
+    depth = argparse.ArgumentParser(add_help=False)
+    depth.add_argument('-d', '--depth', type=common.positive_int,
+                       default=None, metavar="NUM",
+                       help="limit the number of dependency levels")
+    options = argparse.ArgumentParser(add_help=False)
+    options.add_argument('-f', '--force', action='store_true',
+                         help="overwrite uncommitted changes in dependencies")
+    options.add_argument('-c', '--clean', action='store_true',
+                         help="keep ignored files in dependencies")
+    shared = {'formatter_class': common.WideHelpFormatter}
 
     # Main parser
     parser = argparse.ArgumentParser(prog=CLI, description=DESCRIPTION,
-                                     **shared)
+                                     parents=[debug, project], **shared)
 
     subs = parser.add_subparsers(help="", dest='command', metavar="<command>")
 
     # Install parser
     info = "get the specified versions of all dependencies"
     sub = subs.add_parser('install', description=info.capitalize() + '.',
-                          help=info, **shared)
+                          help=info, parents=[debug, project, depth, options],
+                          **shared)
     sub.add_argument('name', nargs='*',
                      help="list of dependencies (`dir` values) to install")
-    sub.add_argument('-f', '--force', action='store_true',
-                     help="overwrite uncommitted changes in dependencies")
-    sub.add_argument('-c', '--clean', action='store_true',
-                     help="keep ignored files in dependencies")
 
     # Update parser
     info = "update dependencies to the latest versions"
     sub = subs.add_parser('update', description=info.capitalize() + '.',
-                          help=info, **shared)
+                          help=info, parents=[debug, project, depth, options],
+                          **shared)
     sub.add_argument('name', nargs='*',
                      help="list of dependencies (`dir` values) to update")
-    # TODO: share force and clean with 'install'
-    sub.add_argument('-f', '--force', action='store_true',
-                     help="overwrite uncommitted changes in dependencies")
-    sub.add_argument('-c', '--clean', action='store_true',
-                     help="keep ignored files in dependencies")
     sub.add_argument('-a', '--all', action='store_true', dest='recurse',
                      help="update all nested dependencies, recursively")
     sub.add_argument('-L', '--no-lock',
@@ -67,7 +68,7 @@ def main(args=None, function=None):
     # Display parser
     info = "display the current version of each dependency"
     sub = subs.add_parser('list', description=info.capitalize() + '.',
-                          help=info, **shared)
+                          help=info, parents=[debug, project, depth], **shared)
     sub.add_argument('-D', '--no-dirty', action='store_false',
                      dest='allow_dirty',
                      help="fail if a source has uncommitted changes")
@@ -75,7 +76,7 @@ def main(args=None, function=None):
     # Uninstall parser
     info = "delete all installed dependencies"
     sub = subs.add_parser('uninstall', description=info.capitalize() + '.',
-                          help=info, **shared)
+                          help=info, parents=[debug, project], **shared)
     sub.add_argument('-f', '--force', action='store_true',
                      help="delete uncommitted changes in dependencies")
 
@@ -101,13 +102,17 @@ def _get_command(function, namespace):
     if namespace.command in ('install', 'update'):
         function = getattr(commands, namespace.command)
         args = namespace.name
-        kwargs.update(force=namespace.force, clean=namespace.clean)
+        kwargs.update(depth=namespace.depth,
+                      force=namespace.force,
+                      clean=namespace.clean)
         if namespace.command == 'update':
-            kwargs.update(recurse=namespace.recurse, lock=namespace.lock)
+            kwargs.update(recurse=namespace.recurse,
+                          lock=namespace.lock)
         exit_msg = "\n" + "Run again with '--force' to overwrite"
     elif namespace.command == 'list':
         function = commands.display
-        kwargs.update(dict(allow_dirty=namespace.allow_dirty))
+        kwargs.update(dict(depth=namespace.depth,
+                           allow_dirty=namespace.allow_dirty))
     elif namespace.command == 'uninstall':
         function = commands.delete
         kwargs.update(force=namespace.force)